# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JornadaMS API
  version: 1.0.0-mvp1
  description: API REST privada do MVP-1 do JornadaMS.
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/v1/time-events:
    post:
      tags: [Attendance]
      operationId: createTimeEvent
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEventCreate'
      responses:
        '201':
          description: Evento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEventResult'
        '200':
          description: Resultado original retornado para reenvio idempotente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEventResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Attendance]
      operationId: listTimeEvents
      x-mvp: MVP-1
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: employee_id
          required: false
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Eventos do período
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEventPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/attendance/days:
    get:
      tags: [Attendance]
      operationId: listDailySummaries
      x-mvp: MVP-1
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: employee_id
          required: false
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Resumos diários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySummaryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: jornada_access
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Conflito de estado ou idempotência
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Dados semanticamente inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    TimeEventCreate:
      type: object
      required: [event_type, occurred_at, source]
      properties:
        event_type:
          type: string
          enum: [ENTRADA, SAIDA]
        occurred_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [WEB]
    TimeEvent:
      allOf:
        - $ref: '#/components/schemas/TimeEventCreate'
        - type: object
          required: [id, employee_id, work_date, status]
          properties:
            id:
              type: string
              format: uuid
            employee_id:
              type: string
              format: uuid
            work_date:
              type: string
              format: date
            status:
              type: string
              enum: [VALID, INVALID, REJECTED]
    DailySummary:
      type: object
      required: [work_date, scheduled_minutes, worked_minutes, balance_minutes, status]
      properties:
        work_date:
          type: string
          format: date
        scheduled_minutes:
          type: integer
        worked_minutes:
          type: integer
        balance_minutes:
          type: integer
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETE, INCONSISTENT]
    TimeEventResult:
      type: object
      required: [event, daily_summary]
      properties:
        event:
          $ref: '#/components/schemas/TimeEvent'
        daily_summary:
          $ref: '#/components/schemas/DailySummary'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
...
    TimeEventPage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeEvent'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    DailySummaryPage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DailySummary'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer